+++ cli/src/execute_ctx.rs
@@ -100,6 +100,7 @@ pub(crate) async fn create_execution_context(
     let object_stores = config.object_stores();
     let secret_stores = config.secret_stores();
     let shielding_sites = config.shielding_sites();
+    let fake_valid_fastly_keys = config.fake_valid_fastly_keys();
     let backend_names = itertools::join(backends.keys(), ", ");
 
     let ctx = ctx
@@ -111,6 +112,7 @@ pub(crate) async fn create_execution_context(
         .with_object_stores(object_stores.clone())
         .with_secret_stores(secret_stores.clone())
         .with_shielding_sites(shielding_sites.clone())
+        .with_fake_valid_fastly_keys(fake_valid_fastly_keys.clone())
         .with_config_path(config_path.into())
         .finish()?;