+++ src/config.rs
@@ -7,7 +7,14 @@ use {
     },
     crate::error::FastlyConfigError,
     serde_derive::Deserialize,
-    std::{collections::HashMap, convert::TryInto, fs, path::Path, str::FromStr, sync::Arc},
+    std::{
+        collections::{HashMap, HashSet},
+        convert::TryInto,
+        fs,
+        path::Path,
+        str::FromStr,
+        sync::Arc,
+    },
     toml::value::Table,
 };
 
@@ -56,6 +63,12 @@ mod secret_store;
 pub use crate::secret_store::{SecretStore, SecretStores};
 
 pub use crate::shielding_site::ShieldingSites;
+
+/// A set of fake valid Fastly keys for testing `fastly_key_is_valid`.
+///
+/// Real Fastly API keys should never be used in local testing. These are fake
+/// values used solely to exercise the `fastly_key_is_valid` hostcall in Viceroy.
+pub type FakeValidFastlyKeys = HashSet<String>;
 
 /// Fastly-specific configuration information.
 ///
@@ -127,6 +140,11 @@ impl FastlyConfig {
     /// Get the shielding site configuration.
     pub fn shielding_sites(&self) -> &ShieldingSites {
         &self.local_server.shielding_sites
+    }
+
+    /// Get the valid mock Fastly API keys configuration.
+    pub fn fake_valid_fastly_keys(&self) -> &FakeValidFastlyKeys {
+        &self.local_server.fake_valid_fastly_keys
     }
 
     /// Parse a `fastly.toml` file into a `FastlyConfig`.
@@ -229,6 +247,7 @@ pub struct LocalServerConfig {
     object_stores: ObjectStoreConfig,
     secret_stores: SecretStoreConfig,
     shielding_sites: ShieldingSites,
+    fake_valid_fastly_keys: FakeValidFastlyKeys,
 }
 
 /// Enum of available (experimental) wasi modules
@@ -253,6 +272,7 @@ struct RawLocalServerConfig {
     object_stores: Option<Table>,
     secret_stores: Option<Table>,
     shielding_sites: Option<Table>,
+    fake_valid_fastly_keys: Option<Vec<String>>,
 }
 
 impl TryInto<LocalServerConfig> for RawLocalServerConfig {
@@ -267,6 +287,7 @@ impl TryInto<LocalServerConfig> for RawLocalServerConfig {
             object_stores,
             secret_stores,
             shielding_sites,
+            fake_valid_fastly_keys,
         } = self;
         let acls = if let Some(acls) = acls {
             acls.try_into()?
@@ -309,6 +330,11 @@ impl TryInto<LocalServerConfig> for RawLocalServerConfig {
             ShieldingSites::default()
         };
 
+        let fake_valid_fastly_keys = fake_valid_fastly_keys
+            .unwrap_or_default()
+            .into_iter()
+            .collect::<FakeValidFastlyKeys>();
+
         Ok(LocalServerConfig {
             acls,
             backends,
@@ -318,6 +344,7 @@ impl TryInto<LocalServerConfig> for RawLocalServerConfig {
             object_stores,
             secret_stores,
             shielding_sites,
+            fake_valid_fastly_keys,
         })
     }
 }